src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/all-in-shadow-tree-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: element.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
